add a column in table and edit vue

+++ database/migrations/2021_01_05_004318_edit_category_name_task_table.php
@@ -25,6 +25,10 @@ class EditCategoryNameTaskTable extends Migration
      */
     public function down()
     {
-        $table->dropColumn('category_name');
+        Schema::table('task', function (Blueprint $table) {
+            if (Schema::hasColumn('category_name')) {
+                $table->dropColumn('category_name');
+            }
+        });
     }
 }
